SPEAKEASY-126: Allow creation of a repo from a push
* If a git repo is requested, has a valid key, and is a valid author requesting it,
  we lazily create the repository
* Worked around ugly JIRA issue where it modifies all content types to include encoding,
  which breaks jgit's smart http detection (jgit Bug 358043)
* Upgraded to jgit 1.0

+++ plugin/src/main/java/com/atlassian/labs/speakeasy/git/DefaultGitRepositoryManager.java
@@ -190,7 +190,7 @@ public class DefaultGitRepositoryManager implements DisposableBean, GitRepositor
         {
             git.commit().
                 setAll(true).
-                setAuthor(event.getUserName(), event.getUserEmail()). //"speakeasy", "speakeasy@atlassian.com").
+                setAuthor(event.getUserName(), event.getUserEmail()).
                 setCommitter("speakeasy", "speakeasy@atlassian.com").
                 setMessage(event.getMessage()).
                 call();
@@ -246,10 +246,13 @@ public class DefaultGitRepositoryManager implements DisposableBean, GitRepositor
             public File operateOn(Repository repo) throws Exception
             {
                 Git git = new Git(repo);
-                git.reset().setMode(ResetCommand.ResetType.HARD).setRef("HEAD").call();
-                for (String path : git.status().call().getUntracked())
+                if (repo.getAllRefs().containsKey("refs/heads/master"))
                 {
-                    new File(repo.getWorkTree(), path).delete();
+                    git.reset().setMode(ResetCommand.ResetType.HARD).setRef("HEAD").call();
+                    for (String path : git.status().call().getUntracked())
+                    {
+                        new File(repo.getWorkTree(), path).delete();
+                    }
                 }
                 return ZipWriter.addDirectoryContentsToJar(repo.getWorkTree(), ".git");
             }
